+++ lib/src/constances/enumeration.dart
@@ -1,3 +1,5 @@
+import 'package:karee/annotations.dart';
+
 ///
 /// KareeInstanceProfile enumeration used to set the profile of the application
 ///
@@ -62,8 +64,22 @@ enum ApplicationKind { cupertino, material }
 /// See [KareeInstanceProfile]('')
 ///
 enum KareeErrorCode {
+  ///
+  ///  Error constant thrown when there is not initial screen in the whole
+  ///  application
+  ///  
   noInitialScreen,
+  ///
+  ///  Error constant thrown when routing is done trough screen name, and the 
+  ///  screen name is not found
+  /// 
+  ///  See [Screen]
+  ///
   screenNotFound,
+  ///
+  /// Error constant that specifies some route is triggered and the path represented 
+  /// by the string didn't exists
+  /// 
   noRouteFound,
   generalError,
   notRoutableWidget,