+++ lib/src/modules/karee_module.dart
@@ -50,12 +50,17 @@ abstract class KareeModule {
   @mustCallSuper
   FutureOr<void> initialize() {
     _isInitialized = true;
-    KareeInternationalization.initAppLocalization()
-        .listen((appLocalization) async {
-      var package = 'packages/$name/';
-      await appLocalization.readModuleTranslationFile(
-          appLocalization.locale!, package);
-    });
+    
+  
+    var package = 'packages/$name/';
+    var obsAppLocalization = KareeInternationalization.initAppLocalization();
+        obsAppLocalization.listen((appLocalization) async =>
+            await appLocalization.readModuleTranslationFile(
+              appLocalization.locale!, package));
+    return obsAppLocalization
+      .value
+      .readModuleTranslationFile(obsAppLocalization.value.locale!, package);
+    
   }
 }